Move the namesapce.go into client.go
