--- k8s/cmd/commands/util/namespace.go
@@ -1,39 +1,0 @@
-/*
-* Copyright 2020-2023 Alibaba Group Holding Limited.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-	http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-package util
-
-import (
-	"github.com/v6d-io/v6d/k8s/cmd/commands/flags"
-	"github.com/v6d-io/v6d/k8s/pkg/log"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-)
-
-// CreateNamespaceIfNotExist creates namespace if it does not exist
-func CreateNamespaceIfNotExist(c client.Client) {
-	namespace := corev1.Namespace{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: flags.Namespace,
-		},
-	}
-
-	if flags.CreateNamespace {
-		if err := CreateIfNotExists(c, &namespace); err != nil {
-			log.Fatal(err, "Expects no positional arguments")
-		}
-	}
-}
